zava-creative-writer-local/src/api/requirements.txt
# Foundry Local SDK
foundry-local-sdk>=0.1.0,<1.0.0
# Web framework
fastapi>=0.100.0,<1.0.0
uvicorn>=0.20.0,<1.0.0
# OpenAI-compatible client
openai>=1.0.0,<2.0.0
# Utilities
pydantic>=2.0.0,<3.0.0
